+++ PhotoGallery.Infrastructure/Migrations/20260112131225_AddThumbnailToPhoto.cs
@@ -1,0 +1,29 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace PhotoGallery.Infrastructure.Migrations
+{
+    /// <inheritdoc />
+    public partial class AddThumbnailToPhoto : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<string>(
+                name: "ThumbnailStorageKey",
+                table: "Photos",
+                type: "nvarchar(max)",
+                nullable: false,
+                defaultValue: "");
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "ThumbnailStorageKey",
+                table: "Photos");
+        }
+    }
+}